Rename pycvrplib to cvrplib
