+++ .pre-commit-config.yaml
@@ -4,20 +4,20 @@ repos:
     hooks:
       - id: system
         name: Black
-        entry: poetry run black pycvrplib
+        entry: poetry run black cvrplib
         pass_filenames: false
         language: system
   - repo: local
     hooks:
       - id: system
         name: isort
-        entry: poetry run isort pycvrplib
+        entry: poetry run isort cvrplib
         pass_filenames: false
         language: system
   - repo: local
     hooks:
       - id: system
         name: MyPy
-        entry: poetry run mypy pycvrplib
+        entry: poetry run mypy cvrplib
         pass_filenames: false
         language: system
