+++ cvrplib/constants.py
