+++ cvrplib/read.py
